# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Resume Analyzer",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/dashboard/student": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Student Dashboard",
        "description": "Return aggregated dashboard data for the currently authenticated student.\nComputes metrics live from the database rather than relying on a\npre-aggregated metrics table.",
        "operationId": "get_student_dashboard_api_dashboard_student_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/dashboard/mentor": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Mentor Dashboard",
        "description": "Return dashboard data for a mentor \u2014 lists all students and their status.",
        "operationId": "get_mentor_dashboard_api_dashboard_mentor_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/dashboard/admin": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Admin Dashboard",
        "description": "Return platform-wide aggregate statistics for admins.",
        "operationId": "get_admin_dashboard_api_dashboard_admin_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/feedback/": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Submit Feedback",
        "description": "Submit mentor feedback for an analysis result.",
        "operationId": "submit_feedback_api_feedback__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/feedback/{analysis_id}": {
      "get": {
        "tags": [
          "feedback"
        ],
        "summary": "Get Feedback For Analysis",
        "description": "Return all mentor feedback entries for a given analysis result.",
        "operationId": "get_feedback_for_analysis_api_feedback__analysis_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedbackResponse"
                  },
                  "title": "Response Get Feedback For Analysis Api Feedback  Analysis Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeedbackCreate": {
        "properties": {
          "analysis_result_id": {
            "type": "string",
            "title": "Analysis Result Id"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "improvement_actions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Improvement Actions"
          }
        },
        "type": "object",
        "required": [
          "analysis_result_id",
          "rating"
        ],
        "title": "FeedbackCreate"
      },
      "FeedbackResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "analysis_result_id": {
            "type": "string",
            "title": "Analysis Result Id"
          },
          "mentor_id": {
            "type": "string",
            "title": "Mentor Id"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "improvement_actions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Improvement Actions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "analysis_result_id",
          "mentor_id",
          "rating",
          "comments",
          "improvement_actions",
          "created_at"
        ],
        "title": "FeedbackResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}